refactor: replace useDeployedTraderSafes composable with getTraderSafes utility function
